Feature: improved compilation of ingredients

+++ src/Entity/RecipeIngredients.php
@@ -52,12 +52,12 @@ class RecipeIngredients
         return $this->id;
     }
 
-    public function getQuantity(): ?int
+    public function getQuantity(): ?float
     {
         return $this->quantity;
     }
 
-    public function setQuantity(int $quantity): self
+    public function setQuantity(float $quantity): self
     {
         $this->quantity = $quantity;
